--- apps/tests/serializers/test_question_crud_serializers.py
@@ -1,79 +1,0 @@
-from rest_framework import serializers
-
-from apps.tests.models import Test, TestQuestion
-
-
-# 등록
-class TestQuestionCreateSerializer(serializers.ModelSerializer):  # type: ignore
-    test_id = serializers.IntegerField(write_only=True)
-    prompt = serializers.CharField(required=False, allow_blank=True)
-    blank_count = serializers.IntegerField(required=False, allow_null=True)
-    options_json = serializers.ListField(child=serializers.CharField(), required=False)
-    answer = serializers.ListField(child=serializers.CharField())
-
-    class Meta:
-        model = TestQuestion
-        fields = [
-            "test_id",
-            "type",
-            "question",
-            "prompt",
-            "blank_count",
-            "options_json",
-            "answer",
-            "point",
-            "explanation",
-        ]
-
-
-# 수정
-class TestQuestionUpdateSerializer(serializers.ModelSerializer):  # type: ignore
-    prompt = serializers.CharField(required=False, allow_blank=True)
-    blank_count = serializers.IntegerField(required=False, allow_null=True)
-    options_json = serializers.ListField(child=serializers.CharField(), required=False)
-    answer = serializers.ListField(child=serializers.CharField())
-
-    class Meta:
-        model = TestQuestion
-        fields = [
-            "type",
-            "question",
-            "prompt",
-            "blank_count",
-            "options_json",
-            "answer",
-            "point",
-            "explanation",
-        ]
-
-
-# 목록 조회
-class TestListItemSerializer(serializers.ModelSerializer):  # type: ignore
-    test_id = serializers.IntegerField(source="id")
-    test_title = serializers.CharField(source="title")
-    subject_title = serializers.CharField(source="subject.title")
-    question_count = serializers.IntegerField()
-    total_score = serializers.IntegerField()
-
-    submission_status = serializers.ChoiceField(choices=["submitted", "not_submitted"])
-    score = serializers.IntegerField(required=False, allow_null=True)
-    correct_count = serializers.IntegerField(required=False, allow_null=True)
-
-    class Meta:
-        model = Test
-        fields = [
-            "test_id",
-            "test_title",
-            "thumbnail_img_url",
-            "subject_title",
-            "question_count",
-            "total_score",
-            "submission_status",
-            "score",
-            "correct_count",
-        ]
-
-
-# 삭제 응답
-class ErrorResponseSerializer(serializers.Serializer):  # type: ignore
-    detail = serializers.CharField()
