+++ config/urls.py
@@ -11,6 +11,7 @@ urlpatterns: list[URLPattern | URLResolver] = [
     path("api/v1/community/", include("apps.community.urls")),
     path("api/v1/admin/", include("apps.courses.urls")),
     path("api/v1/auth/", include("apps.users.urls")),
+    path("api/v1/", include("apps.tests.urls", namespace="tests")),
 ]
 
 if settings.DEBUG:
